pyproject.toml
[project]
name = "greedyqm"
description = "Greedy construction of quadratic manifolds"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "absl-py",
  "numpy",
  "jaxlib",
  "jax"
]

dynamic = ["version"]


[project.optional-dependencies]
dev = [
  "pylint>=2.6.0",
  "yapf"
]

[tool.setuptools]
py-modules = ["greedyqm"]
